Rename job and documentation
